requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.13
distlib==0.3.7
dnspython==2.4.0
filelock==3.12.2
Flask==2.3.2
Flask-API==3.1
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.239
langsmith==0.0.14
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.2
platformdirs==3.9.1
pydantic==1.10.11
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.1
Werkzeug==2.3.6
yarl==1.9.2
